requirements.txt
fastapi>=0.95.0
langchain>=1.0.0
langchain-community>=0.4.0
langchain-openai>=0.1.2
gradio>=4.0.0
python-dotenv>=1.0.0
openai>=1.0.0